* Made category required on all Items.

+++ trunk/epic/categories/models.py
@@ -1,6 +1,8 @@
 from django.contrib.auth.models import User
 from django.db import models
 
+from epic.categories.constants import NO_CATEGORY
+
 
 class Category(models.Model):
     name = models.TextField()
@@ -14,5 +16,16 @@ class Category(models.Model):
         return '%s' % self.name
     
     def delete(self):
-        self.item_set.clear()
+        if self.name == NO_CATEGORY:
+            raise CannotDeleteNoCategoryException()
+        
+        no_category = Category.objects.get(name=NO_CATEGORY)
+        self.item_set.update(category=no_category)
         super(Category, self).delete()
+
+class CannotDeleteNoCategoryException(Exception):
+    def __init__(self):
+        self.value = 'You may not delete the "%s" category.' % NO_CATEGORY
+
+    def __str__(self):
+        return repr(self.value)
